+++ src/main/kotlin/omok/model/entity/StoneColor.kt
@@ -1,4 +1,4 @@
-package omok
+package omok.model.entity
 
 enum class StoneColor {
     BLACK,
